refactor: trie cache configuration (#7566)
This is a pure refactor, including two main refactors:

- replace `TrieCacheFactory` with `TrieConfig`
- move config constants into `TrieConfig`

The motivation is to facilitate wider configuration and changes to the
config format in follow-up PRs.

See also #7564.

+++ core/store/src/config.rs
@@ -44,7 +44,7 @@ pub struct StoreConfig {
     /// Default value: ShardUId {version: 1, shard_id: 3} -> 45_000_000
     /// We're still experimenting with this parameter and it seems decreasing its value can improve
     /// the performance of the storage
-    pub trie_cache_capacities: Vec<(ShardUId, usize)>,
+    pub trie_cache_capacities: Vec<(ShardUId, u64)>,
 }
 
 /// Mode in which to open the storage.
